+++ plugins/cm4all-wp-impex/inc/class-impex-import-rest-controller.php
@@ -486,8 +486,8 @@ class ImpexImportRESTController extends \WP_REST_Controller implements ImpexRest
 }
 
 \add_action(
-  hook_name: 'rest_api_init',
-  callback: function () {
+  'rest_api_init',
+  function () {
     $controller = new ImpexImportRESTController();
     $controller->register_routes();
   },